fix: use my fork of gx.nvim to fix Windows link opening
Special characters don't work with the `powershell.exe start
explorer.exe` chain of commands, but it works when removing the
`explorer.exe` argument. However, there is an upstream bug preventing
Windows users from fully controlling the arguments passed when gx.nvim
opens a link, so I forked it and fixed it.

There is a note in the code change for this commit that mentions we are
waiting on a PR