fix: english short day name increment/decrement support
The previous implementation assumed that the day name was related to a
"date", which is not always the case. Dial.nvim didn't like short day
names that were disconnected in this way, so we define a new constant
augent that is just cyclic between all the short day names.

There may be a better way to do this, because the short day names are
actually defined internally within dial.nvim code; I just don't know
whether a cycling constant augend is defined for them and how to expose
it in the right places. It's defined for the full day names, because
it's working, but I can't find the code!